@@ -17,15 +17,22 @@
  * along with Tuleap. If not, see <http://www.gnu.org/licenses/>.
  */
 
-const path = require("path");
-const { webpack_configurator } = require("@tuleap/build-system-configurator");
+import path from "node:path";
+import {fileURLToPath} from "node:url";
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
 const context = __dirname;
+
+import { webpack_configurator } from "@tuleap/build-system-configurator";
+import POGettextPlugin from "@tuleap/po-gettext-plugin";
+import { VueLoaderPlugin } from "vue-loader";
 const output = webpack_configurator.configureOutput(
     path.resolve(__dirname, "./frontend-assets"),
     "/assets/roadmap/configure-widget/"
 );
 
-module.exports = [
+export default [
     {
         entry: {
             "configure-roadmap-widget-script": "./src/index.ts",
@@ -44,7 +51,6 @@ module.exports = [
                 ...webpack_configurator.configureTypescriptRules(
                     webpack_configurator.babel_options_chrome_firefox
                 ),
-                webpack_configurator.rule_easygettext_loader,
                 webpack_configurator.rule_vue_loader,
                 webpack_configurator.rule_scss_loader,
                 webpack_configurator.rule_css_assets,
@@ -53,11 +59,9 @@ module.exports = [
         plugins: [
             webpack_configurator.getCleanWebpackPlugin(),
             webpack_configurator.getManifestPlugin(),
-            webpack_configurator.getVueLoaderPlugin(),
+            new VueLoaderPlugin(),
+            POGettextPlugin.webpack(),
             ...webpack_configurator.getCSSExtractionPlugins(),
         ],
-        resolveLoader: {
-            alias: webpack_configurator.easygettext_loader_alias,
-        },
     },
 ];
