@@ -1,5 +1,5 @@
-/**
- * Copyright (c) Enalean, 2021 - present. All Rights Reserved.
+/*
+ * Copyright (c) Enalean, 2021 - Present. All Rights Reserved.
  *
  * This file is a part of Tuleap.
  *
@@ -15,18 +15,9 @@
  *
  * You should have received a copy of the GNU General Public License
  * along with Tuleap. If not, see <http://www.gnu.org/licenses/>.
- *
  */
 
-import type { Vue } from "vue/types/vue";
-import { createLocalVue } from "@vue/test-utils";
-import { initVueGettext } from "@tuleap/vue2-gettext-init";
+import common from "./webpack.common.mjs";
+import {webpack_configurator} from "@tuleap/build-system-configurator";
 
-export async function createRoadmapLocalVue(): Promise<typeof Vue> {
-    const local_vue = createLocalVue();
-    await initVueGettext(local_vue, () => {
-        throw new Error("Fallback to default");
-    });
-
-    return local_vue;
-}
+export default webpack_configurator.extendProdConfiguration(common);
